+++ engine/api/rbac/rbac_validation_test.go
@@ -2,6 +2,7 @@ package rbac
 
 import (
 	"context"
+	"fmt"
 	"github.com/ovh/cds/engine/api/group"
 	"github.com/ovh/cds/engine/api/organization"
 	"github.com/ovh/cds/engine/api/region"
@@ -78,4 +79,97 @@ func TestInvalidRbacRegionUsers(t *testing.T) {
 	err := Insert(ctx, db, &r)
 	require.Error(t, err)
 	require.Contains(t, err.Error(), "some users are not part of the allowed organizations")
+}
+
+func TestRBACGlobalInvalidGlobalRole(t *testing.T) {
+	rb := sdk.RBACGlobal{
+		Role:          "runWorkflow",
+		RBACGroupsIDs: []int64{1},
+		RBACUsersIDs:  []string{"aa-aa-aa"},
+	}
+	err := isValidRBACGlobal("myRule", rb)
+	require.Error(t, err)
+	require.Contains(t, err.Error(), "rbac myRule: role runWorkflow is not allowed on a global permission")
+}
+func TestRBACGlobalInvalidGroupAndUsers(t *testing.T) {
+	rb := sdk.RBACGlobal{
+		Role:          sdk.GlobalRoleProjectCreate,
+		RBACGroupsIDs: []int64{},
+		RBACUsersIDs:  []string{},
+	}
+	err := isValidRBACGlobal("myRule", rb)
+	require.Error(t, err)
+	require.Contains(t, err.Error(), "rbac myRule: missing groups or users on global permission")
+}
+
+func TestRBACGlobalEmptyRole(t *testing.T) {
+	rb := sdk.RBACGlobal{
+		Role:          "",
+		RBACGroupsIDs: []int64{1},
+		RBACUsersIDs:  []string{},
+	}
+	err := isValidRBACGlobal("myRule", rb)
+	require.Error(t, err)
+	require.Contains(t, err.Error(), "rbac myRule: role for global permission cannot be empty")
+}
+
+func TestRBACProjectInvalidRole(t *testing.T) {
+	rb := sdk.RBACProject{
+		RBACProjectKeys: []string{"foo"},
+		All:             false,
+		Role:            sdk.GlobalRoleProjectCreate,
+		RBACGroupsIDs:   []int64{1},
+		RBACUsersIDs:    []string{"aa-aa-aa"},
+	}
+	err := isValidRBACProject("myRule", rb)
+	require.Error(t, err)
+	require.Contains(t, err.Error(), fmt.Sprintf("rbac myRule: role %s is not allowed on a project permission", sdk.GlobalRoleProjectCreate))
+}
+func TestRBACProjectInvalidGroupAndUsers(t *testing.T) {
+	rb := sdk.RBACProject{
+		RBACProjectKeys: []string{"foo"},
+		All:             false,
+		Role:            sdk.ProjectRoleRead,
+		RBACGroupsIDs:   []int64{},
+		RBACUsersIDs:    []string{},
+	}
+	err := isValidRBACProject("myRule", rb)
+	require.Error(t, err)
+	require.Contains(t, err.Error(), "rbac myRule: missing groups or users on project permission")
+}
+func TestRBACProjectInvalidProjectKeys(t *testing.T) {
+	rb := sdk.RBACProject{
+		RBACProjectKeys: []string{},
+		All:             false,
+		Role:            sdk.ProjectRoleRead,
+		RBACGroupsIDs:   []int64{1},
+		RBACUsersIDs:    []string{},
+	}
+	err := isValidRBACProject("myRule", rb)
+	require.Error(t, err)
+	require.Contains(t, err.Error(), "rbac myRule: must have at least 1 project on a project permission")
+}
+func TestRBACProjectEmptyRole(t *testing.T) {
+	rb := sdk.RBACProject{
+		RBACProjectKeys: []string{"foo"},
+		All:             false,
+		Role:            "",
+		RBACGroupsIDs:   []int64{1},
+		RBACUsersIDs:    []string{},
+	}
+	err := isValidRBACProject("myRule", rb)
+	require.Error(t, err)
+	require.Contains(t, err.Error(), "rbac myRule: role for project permission cannot be empty")
+}
+func TestRBACProjectInvalidAllAndListOfProject(t *testing.T) {
+	rb := sdk.RBACProject{
+		RBACProjectKeys: []string{"foo"},
+		All:             true,
+		Role:            sdk.ProjectRoleRead,
+		RBACGroupsIDs:   []int64{1},
+		RBACUsersIDs:    []string{},
+	}
+	err := isValidRBACProject("myRule", rb)
+	require.Error(t, err)
+	require.Contains(t, err.Error(), "rbac myRule: you can't have a list of project and the all flag checked on a project permission")
 }
